[auth] refactor openid controller
