+++ apps/boruta_web/lib/boruta_web/router.ex
@@ -53,9 +53,10 @@ defmodule BorutaWeb.Router do
   scope "/openid", BorutaWeb do
     pipe_through(:api)
 
-    get("/jwks", OpenidController, :jwks_index)
-    get("/jwks/:client_id", OpenidController, :jwks_show)
-    post("/register", OpenidController, :register_client)
+    get("/jwks", Openid.JwksController, :jwks_index)
+    get("/jwks/:client_id", Openid.JwksController, :jwks_show)
+
+    post("/register", Openid.DynamicRegistrationController, :register_client)
   end
 
   scope "/oauth", BorutaWeb.Oauth do
@@ -71,8 +72,8 @@ defmodule BorutaWeb.Router do
   scope "/oauth", BorutaWeb do
     pipe_through(:api)
 
-    get("/userinfo", OpenidController, :userinfo)
-    post("/userinfo", OpenidController, :userinfo)
+    get("/userinfo", Openid.UserinfoController, :userinfo)
+    post("/userinfo", Openid.UserinfoController, :userinfo)
   end
 
   scope "/oauth", BorutaWeb.Oauth do
