+++ apps/boruta_web/lib/boruta_web/views/openid_view.ex
@@ -1,6 +1,30 @@
 defmodule BorutaWeb.OpenidView do
   use BorutaWeb, :view
 
+  alias Boruta.Openid.UserinfoResponse
+
+  def render("userinfo.json", %{response: response}) do
+    UserinfoResponse.payload(response)
+  end
+
+  def render("jwks.json", %{keys: keys}) do
+    %{
+      keys: keys
+    }
+  end
+
+  def render("jwk.json", %{client: %Boruta.Ecto.Client{id: client_id, public_key: public_key}}) do
+    {_type, jwk} = public_key |> :jose_jwk.from_pem() |> :jose_jwk.to_map()
+
+    %{
+      keys: [Map.put(jwk, :kid, client_id)]
+    }
+  end
+
+  def render("userinfo.jwt", %{response: response}) do
+    UserinfoResponse.payload(response)
+  end
+
   def render("show.json", %{client: client}) do
     %{data: render_one(client, __MODULE__, "client.json")}
   end
